==> 1-0.Word2Vec & LDA.ipynb <==
# Import

import numpy as np
import pandas as pd
import json
from konlpy.tag import Kkma, Twitter
import re
import gensim

pokr = pd.read_json('./pokr.json')

pokr[:1]

pokr = pokr[['content']] # 발언들만 사용

Tw = Twitter()

# Word2Vec에는 전체 형태소, LDA에는 명사만 사용

Tw.morphs('성원이 되었으므로 제 342회 회의를')

Tw.nouns('성원이 되었으므로 제 342회 회의를')

word2vec_m = []
lda_n = []
for row in pokr.iterrows():
    content = row[1]['content']
    
    real_content = (re.sub('[=……ㆍ◎,◯!―()\n]', '', content))
    
    morphs = Tw.morphs(real_content)
    nouns = Tw.nouns(real_content)
    
    word2vec_m.append(morphs)
    lda_n.append(nouns)

dictionary = corpora.Dictionary(lda_n)

corpus = [dictionary.doc2bow(comment) for comment in lda_n]

word2vecmodel = gensim.models.Word2Vec(all_comment,size=200,min_count=10, sg=1)
ldamodel = gensim.models.ldamodel.LdaModel(corpus = corpus, num_topics = 30, id2word = dictionary, update_every=1, passes=5)

# 결과 확인

word2vecmodel.most_similar(positive=['새누리당', '야당'],negative=['민주당'])

ldamodel.show_topic(1)

word2vecmodel.save('word2vecmodel')
ldamodel.save('ldamodel')
