==> 2.Part_Divide.ipynb <==
import numpy as np
import pandas as pd
import json
from konlpy.tag import Kkma, Twitter
from collections import Counter
import gensim
import re
from gensim import corpora

word2vecmodel = gensim.models.Word2Vec.load('./word2vecmodel')
ldamodel =  gensim.models.ldamodel.LdaModel.load('./ldamodel')
dictionary = corpora.Dictionary.load('ldamodel.id2word')
Tw = Twitter()

with open('con_Q.json', 'r') as fp:
    con_Q= json.load(fp)
with open('con_A.json', 'r') as fp:
    con_A= json.load(fp)
with open('pro_Q.json', 'r') as fp:
    pro_Q= json.load(fp)
with open('pro_A.json', 'r') as fp:
    pro_A= json.load(fp)
    
with open('all_A.json','r') as fp:
    all_A = json.load(fp)
with open('all_Q.json','r') as fp:
    all_Q = json.load(fp)

member = pd.read_json('member.json')

con_Q_part = {}
con_A_part = {}

pro_Q_part = {}
pro_A_part = {}


all_Q_part = {}
all_A_part = {}

con_A_part_name = {}

pro_A_part_name = {}

all_A_part_name = {}

# LDA를 통해 발언들은 주제 별로 part(30개)  나누어서 저장, LDA 확률이 같으면 별도의 31번 째 part에 저장

#모든 파트가 동일한 확률 일 때 check가 0 아니면 1

#part를 정해주는 함수
def checkpart(Q):
    nouns = Tw.nouns(Q)
    
    topics = ldamodel.get_document_topics(dictionary.doc2bow(nouns))

    max_prob = round(topics[0][1], 5)
    temp_prob = round(topics[0][1], 5)
    max_part = topics[0][0]
    
    check = 0

    for i in range(1,len(topics)):
        if round(topics[i][1], 5) != temp_prob:
            check = 1
            break
    
    if check == 0:
        return(30)

    if check == 1:
        for part,prob in topics:
            if round(prob,5) > max_prob:
                max_prob = round(prob,5)
                max_part = part
    
        return(max_part)

for i in range(len(con_Q)):
    Q = con_Q[i][0]
    A = con_A[i][0]
    part = checkpart(Q)
    
    con_Q_part.setdefault(part,[])
    con_Q_part[part].append(Q)
    
    con_A_part.setdefault(part,[])
    con_A_part[part].append(A)
    
for i in range(len(pro_Q)):
    Q = pro_Q[i][0]
    A = pro_A[i][0]
    part = checkpart(Q)
    
    pro_Q_part.setdefault(part,[])
    pro_Q_part[part].append(Q)
    
    pro_A_part.setdefault(part,[])
    pro_A_part[part].append(A)
    
for i in range(len(all_Q)):
    Q = all_Q[i][0]
    A = all_A[i][0]
    part = checkpart(Q)
    
    all_Q_part.setdefault(part,[])
    all_Q_part[part].append(Q)
    
    all_A_part.setdefault(part,[])
    all_A_part[part].append(A)

with open('con_Q_part.json','w') as fp:
    json.dump(con_Q_part,fp)
with open('con_A_part.json','w') as fp:
    json.dump(con_A_part,fp)
with open('pro_Q_part.json','w') as fp:
    json.dump(pro_Q_part,fp)
with open('pro_A_part.json','w') as fp:
    json.dump(pro_A_part,fp)
    
with open('all_Q_part.json','w') as fp:
    json.dump(all_Q_part, fp)
with open('all_A_part.json','w') as fp:
    json.dump(all_A_part, fp)
