==> 5.Lstsq.ipynb <==
# Least Square Method을 이용해서 Y = A * X 에서 A를 찾아낸다

import numpy as np
import pandas as pd
import json

with open('con_Q_part_vec.json','r') as fp:
    con_Q_part_vec = json.load(fp)
with open('con_A_part_vec.json','r') as fp:
    con_A_part_vec = json.load(fp)
with open('pro_Q_part_vec.json','r') as fp:
    pro_Q_part_vec = json.load(fp)
with open('pro_A_part_vec.json','r') as fp:
    pro_A_part_vec = json.load(fp)

with open('all_Q_part_vec.json','r') as fp:
    all_Q_part_vec = json.load(fp)
with open('all_A_part_vec.json','r') as fp:
    all_A_part_vec = json.load(fp)

A_estimate_con = {}
A_estimate_pro = {}
A_estimate_all = {}

for key in con_A_part_vec.keys():
    con_A_part_vec[key] = np.array(con_A_part_vec[key])
    con_Q_part_vec[key] = np.array(con_Q_part_vec[key])
    
    for k in range(230):
        a = con_A_part_vec[key][:,k]
        w = (np.linalg.lstsq(con_Q_part_vec[key], a)[0])
    
        A_estimate_con.setdefault(key,[])
        A_estimate_con[key].append(w)

    A_estimate_con[key] = np.array(A_estimate_con[key]).T
    
    
for key in pro_A_part_vec.keys():
    pro_A_part_vec[key] = np.array(pro_A_part_vec[key])
    pro_Q_part_vec[key] = np.array(pro_Q_part_vec[key])
    
    for k in range(230):
        a = pro_A_part_vec[key][:,k]
        w = (np.linalg.lstsq(pro_Q_part_vec[key], a)[0])
    
        A_estimate_pro.setdefault(key,[])
        A_estimate_pro[key].append(w)

    A_estimate_pro[key] = np.array(A_estimate_pro[key]).T

    
    
for key in all_A_part_vec.keys():
    all_A_part_vec[key] = np.array(all_A_part_vec[key])
    all_Q_part_vec[key] = np.array(all_Q_part_vec[key])
    
    for k in range(230):
        a = all_A_part_vec[key][:,k]
        w = (np.linalg.lstsq(all_Q_part_vec[key], a)[0])
    
        A_estimate_all.setdefault(key,[])
        A_estimate_all[key].append(w)

    A_estimate_all[key] = np.array(A_estimate_all[key]).T

for key in A_estimate_con.keys():
    A_estimate_con[key] = A_estimate_con[key].tolist()
    A_estimate_pro[key] = A_estimate_pro[key].tolist()
    A_estimate_all[key] = A_estimate_all[key].tolist()

with open('./A_estimate_con.json','w') as fp:
    json.dump(A_estimate_con,fp)
with open('./A_estimate_pro.json','w') as fp:
    json.dump(A_estimate_pro,fp)
with open('./A_estimate_all.json','w') as fp:
    json.dump(A_estimate_all,fp)
